Support regressions in eyeD3…, support covers in external files

+++ rename_music.rb
@@ -47,6 +47,23 @@ def rename(options,file,tags)
     end
     puts "\e\[32mmv #{file} #{target_dir}/#{target_file}\e[0m" if options[:verbose]
   end
+
+  if not File.exists? target_dir+"/cover.jpg" and not File.exists? target_dir+"/cover.png" then
+    ext = file.split('.')[-1]
+    if ext == "mp3"
+      if not options[:try]
+        `eyeD3 --write-images #{Shellwords.escape(target_dir)}/ #{Shellwords.escape(file)}`
+        File.rename target_dir+"/FRONT_COVER.jpeg",target_dir+"/cover.jpg"
+      end
+      puts "\e\[32meyeD3 --write-images #{target_dir}/ #{file}" if options[:verbose]
+      puts "\e\[32mmv #{target_dir}/FRONT_COVER.jpeg #{target_dir}/cover.jpg" if options[:verbose]
+    elsif ext == "flac"
+      if not options[:try]
+        `metaflac --export-picture-to=#{Shellwords.escape(target_dir)}/cover.png #{Shellwords.escape(file)}`
+      end
+      puts "\e\[32mmetaflac --export-picture-to=#{target_dir}/cover.png #{file}" if options[:verbose]
+    end
+  end
 end
 
 begin
@@ -84,13 +101,13 @@ begin
       ext = file.split('.')[-1]
       if ext == "mp3"
         tags = {}
-        mp3 = `eyeD3 --rfc822 --no-color #{Shellwords.escape(file)} 2> /dev/null`
+        mp3 = `eyeD3 #{Shellwords.escape(file)} 2> /dev/null`
         raise "#{Shellwords.escape(file)} is not a correct file" if $? != 0
         mp3.each_line do |tag|
-          data = tag.scan /^([A-z ]+): (.*)/
+          data = tag.scan /^\e\[1m(.*)\e\[22m: ([^\n\t]*)/
           key = data[0][0].downcase if not data[0].nil?
           if not data.empty?
-            if key == "year" then tags["date"] = data[0][1]
+            if key == "recording date" then tags["date"] = data[0][1]
             elsif key == "track" then tags["tracknumber"] = data[0][1]
             else tags[key] = data[0][1]
             end
@@ -107,7 +124,7 @@ begin
           tags[type.shift.downcase] = type.join("=")[0..-2]
         end
         rename(options,file,tags)
-      else
+      elsif File.basename(file) != "cover.jpg" and File.basename(file) != "cover.png"
         if not options[:try]
           File.delete(file)
         end
@@ -116,14 +133,25 @@ begin
     end
   end
 
-  Dir['**/'].reverse_each { |d| 
+  Dir['**/'].reverse_each do |d|
     if Dir.entries(d).size == 2
       if not options[:try]
         Dir.rmdir d
       end
       puts "\e\[31mrmdir #{File.absolute_path(d)}\e[0m" if options[:verbose]
+    else
+      exts = []
+      Dir.glob(d+"*").each do |f|
+        if File.file?(f) then
+          exts.push(f.split('.')[-1])
+        end
+      end
+      if not exts.uniq.include?("mp3") and not exts.uniq.include?("flac") and not exts.empty? then
+        puts "\e\[31mrmdir #{d}\e[0m" if options[:verbose]
+      end
     end
-  }
+  end
+    
 rescue Exception => msg
   puts "\e\[31m#{msg}\e[0m"
 end
